Currently broken: Run Mediawiki installation script
This still needs several arguments to the install script before it will
run successfully.  However, the playbook now includes the necessary PHP
modules.  Note that I changed the password to a variable, which *will
break the playbook* in its current form.

Next: Pull in the password from a vault or other safe place and pass in
the correct arguments to the install script for it to run and create a
wiki.