Add create wiki instance through UI
This is largely a proof of concept commit that will be overwritten
once there's a more formalized wiki addition system.

This requires a script to be on your machine that can add a wiki
to the farm, as in MEDIAWIKI_INSTALL.md, with the passwords
and accounts that are included in that file.  It also forces you
to have set up some domain "<wiki>.otswiki.net" as detailed in
that file.