Add standalone flag, and use to disable extends.
In flask, you can use the Null/Master trick to extend a base template if
and only if not running in standalone mode.  This way, when
librehq-wikis is running as part of core, it gets the header/footer/etc,
but when running in standalone mode, none of that is pulled in.