Use MySQL for MW
MediaWiki recommends MySQL, and @frankduncan used it in his local set
up, so switch to that instead of PostgreSQL.

In theory this should have created and deployed two wikis on the server.
The playbook now runs without errors (happily), but when I navigate to
the base URL I see just the "LocalSettings.php not set up yet" message.
The "test1" and "test2" sub-URLs don't resolve. It may be an Apache
configuration error.